requirements.txt
chromadb
numpy
sentence-transformers
python-dotenv
langchain
langchain-groq